update score count + wall spawn

+++ include/gamestate.h
@@ -3,5 +3,5 @@
 enum GameScreen {MAIN_MENU, STAGE, GAMEOVER};
 struct gamestate {
     GameScreen currentScreen = MAIN_MENU;
-    int score = 0;
+    float score = 0;
 };
